--- src/MOCK_DB_DATA/library-css/css-components.data.ts
@@ -1,9 +1,0 @@
-import { LibraryItem } from 'src/app/views/shared/library/models/library-item.model';
-
-import { Tooltip } from './components/tooltip';
-
-export const CssComponentsData: LibraryItem[] = [
-  ...Tooltip,
-];
-
-
