+++ src/MOCK_DB_DATA/library-items/tooltip/_tooltip.data.ts
@@ -1,0 +1,9 @@
+import { LibraryItem } from 'src/app/views/shared/library/models/library-item.model';
+
+import { Tooltip } from './tooltip';
+
+export const TooltipData: LibraryItem[] = [
+  ...Tooltip
+];
+
+
