+++ src/MOCK_DB_DATA/library-items/tooltip/tooltip.ts
@@ -1,10 +1,10 @@
 import { LibraryItem, LibraryItemType } from 'src/app/views/shared/library/models/library-item.model';
 import { BootstrapComponentsBaseUrl, NgBootstrapComponentsBaseUrl } from '../../bootstrap-base-urls';
 
-export const Template: LibraryItem[] = [{
-  id: 'title',
-  title: 'Title',
-  type: LibraryItemType.html,
+export const Tooltip: LibraryItem[] = [{
+  id: 'tooltip',
+  title: 'Tooltip',
+  type: LibraryItemType.ngBootstrap,
   exampleHtml: getExampleHtml(),
   codeHtml: getCodeHtml(),
   documentationHtml: getDocumentationHtml()
@@ -23,12 +23,27 @@ function getExampleHtml(): string {
  * Return null to remove Code from library-item.
  */
 function getCodeHtml(): string | null {
-  return ``;
+  return `
+<button type="button" class="btn fhi-btn-outline-secondary" ngbTooltip="This is the tooltip.">
+  Hover for tooltip
+</button>`;
 }
 
 /*
  * Return null to remove Code from library-item.
  */
 function getDocumentationHtml(): string | null {
-  return ``;
+  return `
+<p>
+  Bootstrap documentation for
+  <a href="${BootstrapComponentsBaseUrl}/tooltips">Tooltip</a>
+</p>
+<p>
+  The Tooltip is implemented as a
+  <a href="${NgBootstrapComponentsBaseUrl}/tooltip">NgBootstrap Tooltip</a>
+  in the FHI Designsystem.
+</p>
+<p>
+  If you use another Javascript framework, you will have to find a suitable 3. party library.
+</p>`;
 }
