+++ src/app/core/mock-db.service.ts
@@ -9,10 +9,8 @@ import { IconsData } from 'src/MOCK_DB_DATA/library-items/icons/_icons.data';
 import { PaginationData } from 'src/MOCK_DB_DATA/library-items/pagination/_pagination.data';
 import { TableData } from 'src/MOCK_DB_DATA/library-items/table/_table.data';
 import { ToastData } from 'src/MOCK_DB_DATA/library-items/toast/_toast.data';
+import { TooltipData } from 'src/MOCK_DB_DATA/library-items/tooltip/_tooltip.data';
 import { TypographyData } from 'src/MOCK_DB_DATA/library-items/typography/_typography.data';
-
-import { CssComponentsData } from 'src/MOCK_DB_DATA/library-css/css-components.data';
-
 
 @Injectable({
   providedIn: 'root'
@@ -28,9 +26,8 @@ export class MockDbService implements InMemoryDbService {
       PaginationData,
       TableData,
       ToastData,
-      TypographyData,
-
-      CssComponentsData,
+      TooltipData,
+      TypographyData
     };
   }
 }
